feat: profile posts pagination, liked-posts tab, gallery bgType storage
iOS PublicProfileView:
- Cursor-based pagination (limit 30) via fetchUserPostsPaginated in SocialService
- "Beğendikleri" segmented tab on own profile reads userLikes/{uid}/items
- fetchLikedPostsPaginated preserves like order, fetches post docs in batch
- Load-more button at bottom of grid for both tabs

DrawingRecord / GalleryStore:
- Added bgType: CanvasBgType and bgColorHex: String fields (Codable, defaults safe for old records)
- thumbnail() now composites bg color + grid/lined pattern + strokes via UIGraphicsImageRenderer
- GalleryStore.save() forwards bgType + bgColor from CanvasView
- GalleryView shares now use record.bgType and record.bgUIColor

DesignSystem:
- Added UIColor(hex:) convenience init and hexString helper

Bug fixes:
- FeedView.swift: missing VStack closing brace caused all helper views to land in local scope
- HashtagFeedView / FeedService: limit(to: Int64) → limit(to: Int) type mismatch
- StatsView: tuple label inference for sorted map
- NotificationsView: missing import FirebaseFirestore for ListenerRegistration

+++ Vibe/Vibe/Services/FeedService.swift
@@ -36,7 +36,7 @@ class FeedService: ObservableObject {
         hasMoreDiscover = true
         discoverListener = db.collection("posts")
             .order(by: "createdAt", descending: true)
-            .limit(to: Int64(pageSize))
+            .limit(to: pageSize)
             .addSnapshotListener { [weak self] snapshot, _ in
                 guard let self, let snap = snapshot else { return }
                 self.lastDiscoverDoc = snap.documents.last
@@ -57,7 +57,7 @@ class FeedService: ObservableObject {
         db.collection("posts")
             .order(by: "createdAt", descending: true)
             .start(afterDocument: lastDoc)
-            .limit(to: Int64(pageSize))
+            .limit(to: pageSize)
             .getDocuments { [weak self] snapshot, _ in
                 guard let self, let snap = snapshot else {
                     self?.isLoadingMore = false
@@ -92,7 +92,7 @@ class FeedService: ObservableObject {
         db.collection("posts")
             .whereField("userId", in: currentFollowingIds)
             .order(by: "createdAt", descending: true)
-            .limit(to: Int64(pageSize))
+            .limit(to: pageSize)
             .getDocuments { [weak self] snap, _ in
                 guard let self, let snap else { return }
                 self.lastFeedDoc = snap.documents.last
@@ -113,7 +113,7 @@ class FeedService: ObservableObject {
             .whereField("userId", in: currentFollowingIds)
             .order(by: "createdAt", descending: true)
             .start(afterDocument: lastDoc)
-            .limit(to: Int64(pageSize))
+            .limit(to: pageSize)
             .getDocuments { [weak self] snap, _ in
                 guard let self, let snap else {
                     self?.isLoadingMoreFeed = false
